hw1: function complete
Two more programs, accuracy.sh and compare, are implemented to
generate the data for the 'accuracy v.s. iterations' chart on pdf
page 38.

In each loop in the accuracy.sh, the models are retrained and new
answer labels are generated. The compare program will compare the
answer labels with the test_seq.txt then the percentage of correct
lables will be printed on the screen. Collect these numbers then you
will be able to draw the chart in excel.